refactor: 즐겨찾기 등록.해제 api url 변경 및 패키지 디렉토리 이동
